Minor Update: Editor -> KeyboardCameraController zoom with defaultzoom casted as float

+++ HousecatEngine/src/editor/Editor.cpp
@@ -263,7 +263,7 @@ void Editor::KeyboardCameraController(SDL_Event& event) {
 		case SDLK_SPACE:
 			camera.x = defaultCamX;
 			camera.y = defaultCamY;
-			zoom = defaultZoom;
+			zoom = static_cast<float>(defaultZoom);
 			break;
 		case SDLK_w:
 			camera.y += camSpeed;
